+++ app/src/androidTest/java/lijewski/demodb/app/main/MainActivityInstrumentedTest.kt
@@ -1,0 +1,36 @@
+package lijewski.demodb.app.main
+
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.filters.MediumTest
+import junit.framework.TestCase.assertEquals
+import lijewski.demodb.app.R
+import lijewski.demodb.presentation.main.MainActivity
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@MediumTest
+@RunWith(AndroidJUnit4::class)
+class MainActivityInstrumentedTest {
+    @get:Rule
+    var activityScenarioRule = ActivityScenarioRule(MainActivity::class.java)
+
+    @Test
+    fun useAppContext() {
+        val appContext = ApplicationProvider.getApplicationContext<Context>()
+
+        assertEquals("lijewski.template", appContext.packageName)
+    }
+
+    @Test
+    fun toolbarIsDisplayed() {
+        onView(withId(R.id.toolbar)).check(matches(isDisplayed()))
+    }
+}
