+++ app/src/main/java/lijewski/demodb/App.kt
@@ -1,0 +1,22 @@
+package lijewski.demodb
+
+import dagger.android.AndroidInjector
+import dagger.android.DaggerApplication
+import dagger.android.HasAndroidInjector
+import lijewski.demodb.di.app.DaggerAppComponent
+import timber.log.Timber
+
+class App : DaggerApplication(), HasAndroidInjector {
+    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
+        return DaggerAppComponent
+            .builder()
+            .application(this)
+            .build()
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+
+        Timber.plant(Timber.DebugTree())
+    }
+}
