+++ app/src/main/java/lijewski/demodb/databinding/CustomBindingAdapter.kt
@@ -1,0 +1,14 @@
+package lijewski.demodb.databinding
+
+import android.view.View
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("visibleGone")
+fun visibleGone(view: View, show: Boolean) {
+    view.visibility = if (show) View.VISIBLE else View.GONE
+}
+
+@BindingAdapter("visibleInvisible")
+fun visibleInvisible(view: View, show: Boolean) {
+    view.visibility = if (show) View.VISIBLE else View.INVISIBLE
+}
