+++ app/src/main/java/lijewski/demodb/di/module/MainModule.kt
@@ -1,0 +1,12 @@
+package lijewski.demodb.di.module
+
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+import lijewski.demodb.presentation.main.MainFragment
+
+@Suppress("unused")
+@Module
+abstract class MainModule {
+    @ContributesAndroidInjector
+    abstract fun contributeMainFragment(): MainFragment
+}
