+++ app/src/main/java/lijewski/demodb/presentation/main/MainFragment.kt
@@ -1,0 +1,49 @@
+package lijewski.demodb.presentation.main
+
+import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
+import dagger.android.support.AndroidSupportInjection
+import lijewski.demodb.app.R
+import lijewski.demodb.app.databinding.MainFragmentBinding
+import javax.inject.Inject
+
+class MainFragment : Fragment() {
+    companion object {
+        const val TAG: String = "MainFragment"
+    }
+
+    @Inject
+    lateinit var viewModelFactory: ViewModelProvider.Factory
+
+    private lateinit var mainViewModel: MainViewModel
+
+    private lateinit var binding: MainFragmentBinding
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        AndroidSupportInjection.inject(this)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        mainViewModel = ViewModelProviders.of(this, viewModelFactory).get(MainViewModel::class.java)
+
+        binding = DataBindingUtil.inflate<MainFragmentBinding>(
+            inflater, R.layout.main_fragment, container, false
+        ).apply {
+            viewModel = mainViewModel
+            lifecycleOwner = this@MainFragment
+        }
+
+        return binding.root
+    }
+}
