+++ app/src/main/java/lijewski/demodb/presentation/splash/SplashActivity.kt
@@ -1,0 +1,21 @@
+package lijewski.demodb.presentation.splash
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import lijewski.demodb.presentation.main.MainActivity
+
+/**
+ * Splash activity is displayed during application's startup
+ */
+class SplashActivity: AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val intent = Intent(this, MainActivity::class.java)
+        startActivity(intent)
+        finish()
+    }
+
+}
