--- app/src/main/java/lijewski/template/App.kt
@@ -1,19 +1,0 @@
-package lijewski.template
-
-import dagger.android.AndroidInjector
-import dagger.android.DaggerApplication
-import dagger.android.HasAndroidInjector
-import lijewski.template.di.app.DaggerAppComponent
-import timber.log.Timber
-
-class App : DaggerApplication(), HasAndroidInjector {
-    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
-        return DaggerAppComponent.factory().create(this)
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-
-        Timber.plant(Timber.DebugTree())
-    }
-}
