--- app/src/main/java/lijewski/template/di/app/AppComponent.kt
@@ -1,22 +1,0 @@
-package lijewski.template.di.app
-
-import dagger.Component
-import dagger.android.AndroidInjector
-import dagger.android.support.AndroidSupportInjectionModule
-import lijewski.template.App
-import lijewski.template.di.module.AppModule
-import lijewski.template.di.module.BuilderModule
-import javax.inject.Singleton
-
-@Singleton
-@Component(
-    modules = [
-        AndroidSupportInjectionModule::class,
-        AppModule::class,
-        BuilderModule::class
-    ]
-)
-interface AppComponent : AndroidInjector<App> {
-    @Component.Factory
-    abstract class Factory : AndroidInjector.Factory<App>
-}
