--- app/src/main/java/lijewski/template/di/module/AppModule.kt
@@ -1,14 +1,0 @@
-package lijewski.template.di.module
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import lijewski.template.App
-
-@Module(includes = [ViewModelModule::class])
-class AppModule {
-    @Provides
-    fun provideContext(application: App): Context {
-        return application.applicationContext
-    }
-}
