--- app/src/main/java/lijewski/template/presentation/splash/SplashActivity.kt
@@ -1,21 +1,0 @@
-package lijewski.template.presentation.splash
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import lijewski.template.presentation.main.MainActivity
-
-/**
- * Splash activity is displayed during application's startup
- */
-class SplashActivity: AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val intent = Intent(this, MainActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-
-}
