+++ app/src/main/res/layout/main_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="lijewski.template.presentation.main.MainActivity">
+    tools:context="lijewski.demodb.presentation.main.MainActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
