+++ app/src/main/res/layout/main_fragment.xml
@@ -9,13 +9,13 @@
 
         <variable
             name="viewModel"
-            type="lijewski.template.presentation.main.MainViewModel" />
+            type="lijewski.demodb.presentation.main.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="lijewski.template.presentation.main.MainFragment">
+        tools:context="lijewski.demodb.presentation.main.MainFragment">
 
         <ProgressBar
             style="?android:attr/progressBarStyleLarge"
