+++ build.gradle
@@ -1,11 +1,10 @@
 apply from: 'buildsystem/dependencies.gradle'
 
-buildscript { 
-    ext.kotlin_version = '1.3.50'
-
+buildscript {
     ext {
-        gradle_version = '3.5.0'
-        kotlin_version = '1.3.50'
+        gradle_version = '3.6.1'
+        kotlin_version = '1.3.61'
+        navigation_version = '2.2.0'
     }
 
     repositories {
@@ -15,6 +14,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:$gradle_version"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$navigation_version"
     }
 }
