+++ buildsystem/dependencies.gradle
@@ -8,15 +8,18 @@ ext {
     material_version = '1.0.0'
     constraintlayout_version = '1.1.3'
 
-    lifecycle_version = '2.1.0'
-    lifecycle_ktx_version = '2.2.0-rc02'
-    navigation_version = '2.0.0'
+    lifecycle_version = '2.2.0'
+    lifecycle_ktx_version = '2.2.0'
 
     timber_version = '4.7.1'
 
-    coroutines_version = '1.3.2'
+    coroutines_version = '1.3.3'
 
     dagger_version = '2.24'
+
+    room_version = '2.2.4'
+
+    gson_version = '2.8.5'
 
     glide_version = '4.8.0'
 
@@ -31,15 +34,22 @@ ext {
     orchestrator_version = '1.2.0'
 
     androidXDependencies = [
-            core      : "androidx.core:core-ktx:$core_ktx_version",
-            appcompat : "androidx.appcompat:appcompat:$appcompat_version",
-            constraint: "androidx.constraintlayout:constraintlayout:$constraintlayout_version",
-            recycler  : "androidx.recyclerview:recyclerview:$recyclerview_version",
-            lifecycle : "androidx.lifecycle:lifecycle-extensions:$lifecycle_version",
-            livedata  : "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_ktx_version",
-            runtime   : "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_ktx_version",
-            viewmodel : "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_ktx_version",
-            material  : "com.google.android.material:material:$material_version"
+            core           : "androidx.core:core-ktx:$core_ktx_version",
+            appcompat      : "androidx.appcompat:appcompat:$appcompat_version",
+            constraint     : "androidx.constraintlayout:constraintlayout:$constraintlayout_version",
+            recycler       : "androidx.recyclerview:recyclerview:$recyclerview_version",
+            lifecycle      : "androidx.lifecycle:lifecycle-extensions:$lifecycle_version",
+            reactivestreams: "androidx.lifecycle:lifecycle-reactivestreams:$lifecycle_version",
+            livedata       : "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_ktx_version",
+            runtime        : "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_ktx_version",
+            viewmodel      : "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_ktx_version",
+            material       : "com.google.android.material:material:$material_version",
+            navigation     : "androidx.navigation:navigation-fragment-ktx:$navigation_version",
+            navigationui   : "androidx.navigation:navigation-ui-ktx:$navigation_version"
+    ]
+
+    androidXKaptDependencies = [
+            compiler: "androidx.lifecycle:lifecycle-compiler:$lifecycle_version"
     ]
 
     daggerDependencies = [
@@ -58,6 +68,16 @@ ext {
             android: "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     ]
 
+    roomDependencies = [
+            runtime : "androidx.room:room-runtime:$room_version",
+            ktx : "androidx.room:room-ktx:$room_version"
+            //coroutines : "androidx.room:room-coroutines:2.1.0-alpha04"
+    ]
+
+    roomKaptDependencies = [
+           roomKapt : "androidx.room:room-compiler:$room_version"
+    ]
+
     loggingDependencies = [
             timber: "com.jakewharton.timber:timber:$timber_version"
     ]
@@ -66,9 +86,8 @@ ext {
             kotlinStdlib: "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     ]
 
-    glideDependencies = [
-            glide    : "com.github.bumptech.glide:glide:$glide_version",
-            glideKapt: "com.github.bumptech.glide:compiler:$glide_version"
+    gsonDependencies = [
+            gson : "com.google.code.gson:gson:$gson_version"
     ]
 
     testDependencies = [
